# bot/src/main/java/backend/academy/bot/model/ApiErrorResponse.java
package backend.academy.bot.model;

import java.util.List;

public record ApiErrorResponse(
    String description,
    String code,
    String exceptionName,
    String exceptionMessage,
    List<String> stacktrace
) {}